+++ src/lib.rs
@@ -188,6 +188,7 @@
 
 pub mod cns;
 pub mod ei_header;
+pub mod frame_reconstruct;
 pub mod framing;
 pub mod huffman;
 pub mod packet_stream;